% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grok.R
\name{grok}
\alias{grok}
\title{grok}
\usage{
grok(
  pattern = "",
  custom_patterns_dir = NULL,
  custom_patterns_list = reticulate::dict()
)
}
\arguments{
\item{pattern}{the pattern to be matched}

\item{custom_patterns_dir}{used to pass custom patterns}

\item{custom_patterns_list}{used to pass custom patterns}
}
\value{
list with variable names specified in pattern and their corresponding values
}
\description{
grok
}
