% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_pygrok}
\alias{install_pygrok}
\title{Install pygrok}
\usage{
install_pygrok(method = "auto", conda = "auto")
}
\arguments{
\item{method}{Installation method, passed to py_install}

\item{conda}{Path to conda, passed to py_install}
}
\description{
Install pygrok
}
\examples{
\dontrun{
install_pygrok()
}
}
