% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{pygrok}
\alias{pygrok}
\title{pygrok module object}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.
}
\usage{
pygrok
}
\description{
Uses the reticulate framework to access the pygrok module
}
\keyword{datasets}
